example overhaul: airline_assets -> dagster_graph_backed_assets 3.1/n
