+++ examples/dagster_graph_backed_assets/README.md
@@ -1,0 +1,57 @@
+# dagster_graph_backed_assets
+
+This is a [Dagster](https://dagster.io/) project scaffolded with [`dagster project scaffold`](https://docs.dagster.io/getting-started/create-new-project).
+
+## Getting started
+
+First, install your Dagster repository as a Python package. By using the --editable flag, pip will install your repository in ["editable mode"](https://pip.pypa.io/en/latest/topics/local-project-installs/#editable-installs) so that as you develop, local code changes will automatically apply.
+
+```bash
+pip install -e .
+```
+
+Then, start the Dagit web server:
+
+```bash
+dagit
+```
+
+Open http://localhost:3000 with your browser to see the project.
+
+You can start writing assets in `dagster_graph_backed_assets/assets/`. The assets are automatically loaded into the Dagster repository as you define them.
+
+## Development
+
+
+### Adding new Python dependencies
+
+You can specify new Python dependencies in `setup.py`.
+
+If you have concrete requirements, such as pins, index selection, or other pip flags needed for a specific environment (e.g. cloud deployment), you can include these in `requirements.txt`.
+
+
+### Unit testing
+
+Tests are in the `dagster_graph_backed_assets_tests` directory and you can run tests using `pytest`:
+
+```bash
+pytest dagster_graph_backed_assets_tests
+```
+
+### Schedules and sensors
+
+If you want to enable Dagster [Schedules](https://docs.dagster.io/concepts/partitions-schedules-sensors/schedules) or [Sensors](https://docs.dagster.io/concepts/partitions-schedules-sensors/sensors) for your jobs, start the [Dagster Daemon](https://docs.dagster.io/deployment/dagster-daemon) process in the same folder as your `workspace.yaml` file, but in a different shell or terminal.
+
+The `$DAGSTER_HOME` environment variable must be set to a directory for the daemon to work. Note: using directories within /tmp may cause issues. See [Dagster Instance default local behavior](https://docs.dagster.io/deployment/dagster-instance#default-local-behavior) for more details.
+
+```bash
+dagster-daemon run
+```
+
+Once your Dagster Daemon is running, you can start turning on schedules and sensors for your jobs.
+
+## Deploy on Dagster Cloud
+
+The easiest way to deploy your Dagster project is to use Dagster Cloud.
+
+Check out the [Dagster Cloud Documentation](https://docs.dagster.cloud) to learn more.
