+++ examples/dagster_graph_backed_assets/dagster_graph_backed_assets/__init__.py
@@ -1,0 +1,1 @@
+from .repository import dagster_graph_backed_assets
