+++ examples/dagster_graph_backed_assets/dagster_graph_backed_assets/assets/__init__.py
@@ -1,0 +1,1 @@
+from .airlines import passenger_flights
