+++ examples/dagster_graph_backed_assets/dagster_graph_backed_assets/repository.py
@@ -1,35 +1,18 @@
+from dagster import load_assets_from_package_module, repository
+
+from dagster_graph_backed_assets import assets
+
 import pandas as pd
 
 from dagster import (
     AssetSelection,
     AssetsDefinition,
     GraphOut,
-    asset,
     define_asset_job,
     graph,
     op,
     repository,
 )
-
-
-@asset
-def passenger_flights():
-    return pd.DataFrame(
-        {
-            "name": ["Harry Potter", "Ron Weasley", "Hermione Granger"],
-            "date": ["2022-07-01", "2022-07-01", "2022-07-01"],
-            "departure_city": ["San Francisco", "Los Angeles", "New York"],
-            "departure_state": ["CA", "CA", "NY"],
-            "departure_country": ["USA", "USA", "USA"],
-            "arrival_city": ["San Francisco", "Los Angeles", "New York"],
-            "arrival_state": ["CA", "CA", "NY"],
-            "arrival_country": ["USA", "USA", "USA"],
-            "age": [20, 21, 39],
-            "rebooked_due_to_cancellation": [False, False, False],
-            "num_layovers": [0, 1, 2],
-            "distance_in_miles": [100, 200, 300],
-        }
-    )
 
 
 @op
@@ -75,10 +58,10 @@ airline_job = define_asset_job("airline_job", AssetSelection.keys("passenger_fli
 
 
 @repository
-def airline_repository():
+def dagster_graph_backed_assets():
     return [
+        load_assets_from_package_module(assets),
         define_asset_job("airline_job", AssetSelection.keys("passenger_flights").downstream()),
-        passenger_flights,
         AssetsDefinition.from_graph(us_assets),
         AssetsDefinition.from_graph(layover_breakdown_2022),
     ]
