+++ examples/dagster_graph_backed_assets/requirements.txt
@@ -1,0 +1,3 @@
+-e dagster_graph_backed_assets
+# It's recommended to include environment-specific requirements in this file.
+# For example, pinning a package's version for deployment.
