+++ examples/dagster_graph_backed_assets/setup.cfg
@@ -1,0 +1,2 @@
+[metadata]
+name = dagster_graph_backed_assets
