+++ examples/dagster_graph_backed_assets/tox.ini
@@ -21,4 +21,4 @@ commands =
 
 [testenv:pylint]
 commands =
-  pylint -j0 --rcfile=../../pyproject.toml {posargs} airline_assets airline_assets_tests
+  pylint -j0 --rcfile=../../pyproject.toml {posargs} dagster_graph_backed_assets dagster_graph_backed_assets_tests
