+++ examples/dagster_graph_backed_assets/workspace.yaml
@@ -1,0 +1,2 @@
+load_from:
+  - python_package: dagster_graph_backed_assets
